feat: add types to photos slice
